Crate massa_metrics

this library is used to collect metrics from the node and expose them to the prometheus server

the metrics are collected from the node and from the survey the survey is a separate thread that is used to collect metrics from the network (active connections)

Modules

server 🔒

Structs

BLOCKS_COUNTER 🔒

DEFERRED_CALLS_TOTAL_GAS 🔒

DEFERRED_CALL_CANCELLED 🔒

DEFERRED_CALL_EXECUTED 🔒

DEFERRED_CALL_FAILED 🔒

DEFERRED_CALL_REGISTERED 🔒

ENDORSEMENTS_COUNTER 🔒

EVENT_CACHE_VEC 🔒

FACTORY_BLOCKS_PRODUCED 🔒

FACTORY_BLOCKS_PRODUCED_WITH_POOL_TIMEOUT 🔒

FACTORY_POOL_TIMEOUT_DENUNCIATIONS 🔒

FACTORY_POOL_TIMEOUT_ENDORSEMENTS 🔒

FACTORY_POOL_TIMEOUT_OPERATIONS 🔒

MassaMetrics

MetricsStopper

OPERATIONS_COUNTER 🔒

Functions

dec_deferred_calls_cancelled_by

dec_deferred_calls_registered_by

inc_deferred_calls_cancelled

inc_deferred_calls_executed_by

inc_deferred_calls_failed_by

inc_deferred_calls_registered

inc_factory_blocks_produced

inc_factory_pool_timeout_denunciations

inc_factory_pool_timeout_endorsements

inc_factory_pool_timeout_operations

register_lazy_static_metrics 🔒

Eagerly register lazy_static metrics so they appear on /metrics even before the first increment (e.g. factory pool timeouts that have not occurred yet).

set_blocks_counter

set_deferred_calls_total_gas

set_endorsements_counter

set_event_cache_vec_len

set_operations_counter