Crate massa_metrics

this library is used to collect metrics from the node and expose them to the prometheus server

the metrics are collected from the node and from the survey the survey is a separate thread that is used to collect metrics from the network (active connections)

Modules

server 🔒

Structs

BLOCKS_COUNTER 🔒

ENDORSEMENTS_COUNTER 🔒

MassaMetrics

MetricsStopper

OPERATIONS_COUNTER 🔒

Functions

set_blocks_counter

set_endorsements_counter

set_operations_counter