Crate massa_metrics

this library is used to collect metrics from the node and expose them to the prometheus server

the metrics are collected from the node and from the survey the survey is a separate thread that is used to collect metrics from the network (active connections)

Modules

server 🔒

survey 🔒

Structs

BLOCKS_COUNTER 🔒

ENDORSEMENTS_COUNTER 🔒

MassaMetrics

OPERATIONS_COUNTER 🔒

Functions

dec_endorsements_counter

dec_operations_counter

inc_endorsements_counter

inc_operations_counter

set_blocks_counter